Rename callable.py to check_callable.py, refs #2078
