+++ datasette/utils/check_callable.py
